+++ server/index.js
@@ -2,7 +2,7 @@ const express = require('express');
 const { genericAnalysis } = require('./azure/genericAnalysis');
 const { analyzeSentiment } = require('./azure/analyzeSentiment');
 const cors = require('cors');
-const analyzeConversations = require('./azure/analyzeConversations');
+const queryKnowledgeBase = require('./azure/queryKnowledgeBase');
 const app = express();
 const port = 3001;
 
@@ -28,9 +28,9 @@ app.get('/sentimentAnalysis', async (req, res) => {
   res.status(200).send(result);
 });
 
-app.get('/conversationsAnalysis', async (req, res) => {
+app.get('/queryKnowledgeBase', async (req, res) => {
   const text = req.query.text;
-  const result = await analyzeConversations(text);
+  const result = await queryKnowledgeBase(text);
 
   res.status(200).send(result);
 });
